+++ protocols/gossipsub/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.33.1 [unreleased]
+
+- Use `instant` and `futures-timer` instead of `wasm-timer` (see [PR 2245]).
+
+[PR 2245]: https://github.com/libp2p/rust-libp2p/pull/2245
+
 # 0.33.0-rc.1 [2021-10-15]
 
 - Add an event to register peers that do not support the gossipsub protocol