+++ protocols/kad/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.32.1 [unreleased]
+
+- Use `instant` and `futures-timer` instead of `wasm-timer` (see [PR 2245]).
+
+[PR 2245]: https://github.com/libp2p/rust-libp2p/pull/2245
+
 # 0.32.0-rc.2 [2021-10-15]
 
 - Export `KademliaBucketInserts` (see [PR 2294]).